Major refactoring that really should have been multiple commits:
Refactored and improved solving algorithm to fix a bug that was discovered.
Added a display step class, a list of which is added concurrently so that if the solver takes a while the solution can begin to display before the solver has finished.
Improved solution display by having it run on an animation timer rather than having the solver update the UI directly (removed bad practice).
Improved solution display UI by implementing an X-mark for crossed cells and getting rid of red box for unknown.
Made use of Streams API, making sure to pass pure functions, or only using it to obtain iterators when mutability is needed.
Split classes out and did a lot of renaming for readability, added classes to prevent code reuse.